# Conflicts:
#	src/Board/Board.java
#	src/Board/MatrixBoard.java

+++ src/State/State.java
@@ -61,7 +61,7 @@ public abstract class State<T, S> {
         return cameFrom;
     }
 
-    public abstract int generateCost();
+    public abstract double generateCost();
 
    /*
    Returns a backTrace of the states for the algorithms
